Library

Library

Thunks module

Jobs module

Workflows module

Thunks module

SimpleWorkflows.Thunks.Thunk — Type

Thunk(::Function, args::Tuple, kwargs::NamedTuple)
Thunk(::Function, args...; kwargs...)
Thunk(::Function)

Hold a Function and its arguments for lazy evaluation. Use reify! to evaluate.

Examples

julia> a = Thunk(x -> 3x, 4);

julia> reify!(a)
Some(12)

julia> b = Thunk(+, 4, 5);

julia> reify!(b)
Some(9)

julia> c = Thunk(sleep)(1);

julia> getresult(c)  # `c` has not been evaluated

julia> reify!(c)  # `c` has been evaluated
Some(nothing)

julia> f(args...; kwargs...) = collect(kwargs);

julia> d = Thunk(f)(1, 2, 3; x=1.0, y=4, z="5");

julia> reify!(d)
Some(Pair{Symbol, Any}[:x => 1.0, :y => 4, :z => "5"])

julia> e = Thunk(sin, "1");  # Catch errors

julia> reify!(e);

SimpleWorkflows.Thunks.ErredResult — Type

Capture errors and stack traces from a running Thunk.

SimpleWorkflows.Thunks.reify! — Function

reify!(thunk::Thunk)

Reify a Thunk into a value.

Compute the value of the expression. Walk through the Thunk's arguments and keywords, recursively evaluating each one, and then evaluating the Thunk's function with the evaluated arguments.

See also Thunk.

SimpleWorkflows.Thunks.getresult — Method

getresult(thunk::Thunk)

Get the result of a Thunk. If thunk has not been evaluated, return nothing, else return a Some-wrapped result.

Jobs module

SimpleWorkflows.Jobs.Job — Type

Job(thunk::Thunk; desc="", user="")

Create a simple job.

Arguments

thunk: a Thunk that encloses the job definition.

desc::String="": describe briefly what this job does.

user::String="": indicate who executes this job.

Examples

julia> a = Job(Thunk(sleep)(5); user="me", desc="Sleep for 5 seconds");

julia> b = Job(Thunk(run, `pwd` & `ls`); user="me", desc="Run some commands");

SimpleWorkflows.Thunks.getresult — Method

getresult(job::Job)

Get the running result of a Job.

The result is wrapped by a Some type. Use something to retrieve its value. If it is nothing, the Job is not finished.

SimpleWorkflows.Jobs.getstatus — Method

getstatus(x::Job)

Get the current status of a Job.

SimpleWorkflows.Jobs.ispending — Function

Test if the Job is still pending.

SimpleWorkflows.Jobs.isrunning — Function

Test if the Job is running.

SimpleWorkflows.Jobs.isexited — Function

Test if the Job has exited.

SimpleWorkflows.Jobs.issucceeded — Function

Test if the Job was successfully run.

SimpleWorkflows.Jobs.isfailed — Function

Test if the Job failed during running.

SimpleWorkflows.Jobs.isinterrupted — Function

Test if the Job was interrupted during running.

SimpleWorkflows.Jobs.pendingjobs — Function

pendingjobs(wf::Workflow)

Filter only the pending jobs in a Workflow.

pendingjobs(jobs)

Filter only the pending jobs in a sequence of Jobs.

SimpleWorkflows.Jobs.runningjobs — Function

runningjobs(wf::Workflow)

Filter only the pending jobs in a Workflow.

runningjobs(jobs)

Filter only the running jobs in a sequence of Jobs.

SimpleWorkflows.Jobs.exitedjobs — Function

exitedjobs(wf::Workflow)

Filter only the pending jobs in a Workflow.

exitedjobs(jobs)

Filter only the exited jobs in a sequence of Jobs.

SimpleWorkflows.Jobs.succeededjobs — Function

succeededjobs(wf::Workflow)

Filter only the pending jobs in a Workflow.

succeededjobs(jobs)

Filter only the succeeded jobs in a sequence of Jobs.

SimpleWorkflows.Jobs.failedjobs — Function

failedjobs(wf::Workflow)

Filter only the pending jobs in a Workflow.

failedjobs(jobs)

Filter only the failed jobs in a sequence of Jobs.

SimpleWorkflows.Jobs.interruptedjobs — Function

interruptedjobs(wf::Workflow)

Filter only the pending jobs in a Workflow.

interruptedjobs(jobs)

Filter only the interrupted jobs in a sequence of Jobs.

SimpleWorkflows.Jobs.createdtime — Function

Return the created time of a Job.

SimpleWorkflows.Jobs.starttime — Function

starttime(job::Job)

Return the start time of a Job. Return nothing if it is still pending.

SimpleWorkflows.Jobs.stoptime — Function

stoptime(job::Job)

Return the stop time of a Job. Return nothing if it has not exited.

SimpleWorkflows.Jobs.elapsed — Function

elapsed(job::Job)

Return the elapsed time of a Job since it started running.

If nothing, the Job is still pending. If it is finished, return how long it took to complete.

SimpleWorkflows.Jobs.description — Function

description(job::Job)

Return the description of a Job.

SimpleWorkflows.run! — Method

run!(job::Job; n=1, δt=1)

Run a Job with maximum n attempts, with each attempt separated by δt seconds.

SimpleWorkflows.Jobs.interrupt! — Function

interrupt!(job::Job)

Manually interrupt a Job, works only if it is running.

SimpleWorkflows.Jobs.queue — Function

queue(; sortby = :created_time)

Print all Jobs that are pending, running, or finished as a table.

Accpetable arguments for sortby are :created_time, :user, :start_time, :stop_time, :elapsed, :status, and :times.

SimpleWorkflows.Jobs.query — Function

query(id::Integer)
query(ids::AbstractVector{<:Integer})

Query a specific (or a list of Jobs) by its (theirs) ID.

SimpleWorkflows.Jobs.ntimes — Function

ntimes(id::Integer)
ntimes(job::Job)

Return how many times a Job has been rerun.

Workflows module

SimpleWorkflows.Workflows.Workflow — Type

Workflow(jobs, graph)

Create a Workflow from a list of Jobs and a graph representing their relations.

SimpleWorkflows.run! — Method

run!(wf::Workflow; n=5, δt=1, Δt=1)

Run a Workflow with maximum n attempts, with each attempt separated by Δt seconds.

Cool down for δt seconds after each Job in the Workflow.

SimpleWorkflows.Jobs.getstatus — Method

getstatus(wf::Workflow)

Get the current status of Jobs in a Workflow.

SimpleWorkflows.Workflows.chain — Function

chain(x::Job, y::Job, z::Job...)

Chain multiple Jobs one after another.

SimpleWorkflows.Workflows.:→ — Function

→(x, y)

Chain two Jobs.

SimpleWorkflows.Workflows.:← — Function

←(y, x)

Chain two Jobs reversely.

SimpleWorkflows.Workflows.thread — Function

thread(xs::AbstractVector{Job}, ys::AbstractVector{Job}, zs::AbstractVector{Job}...)

Chain multiple vectors of Jobs, each Job in xs has a corresponding Job in ys.`

SimpleWorkflows.Workflows.:⇶ — Function

⇶(xs, ys)

Chain two vectors of Jobs.

SimpleWorkflows.Workflows.:⬱ — Function

⬱(ys, xs)

Chain two vectors of Jobs reversely.

SimpleWorkflows.Workflows.fork — Function

fork(x::Job, ys::AbstractVector{Job})
⇉(x, ys)

Attach a group of parallel Jobs (ys) to a single Job (x).

SimpleWorkflows.Workflows.converge — Function

converge(xs::AbstractVector{Job}, y::Job)
⭃(xs, y)

Finish a group a parallel Jobs (xs) by a single Job (y).

SimpleWorkflows.Workflows.spindle — Function

spindle(x::Job, ys::AbstractVector{Job}, z::Job)

Start from a Job (x), followed by a series of Jobs (ys), finished by a single Job (z).

SimpleWorkflows.Jobs.pendingjobs — Method

pendingjobs(wf::Workflow)

Filter only the pending jobs in a Workflow.

SimpleWorkflows.Jobs.runningjobs — Method

runningjobs(wf::Workflow)

Filter only the pending jobs in a Workflow.

SimpleWorkflows.Jobs.exitedjobs — Method

exitedjobs(wf::Workflow)

Filter only the pending jobs in a Workflow.

SimpleWorkflows.Jobs.succeededjobs — Method

succeededjobs(wf::Workflow)

Filter only the pending jobs in a Workflow.

SimpleWorkflows.Jobs.failedjobs — Method

failedjobs(wf::Workflow)

Filter only the pending jobs in a Workflow.

SimpleWorkflows.Jobs.interruptedjobs — Method

interruptedjobs(wf::Workflow)

Filter only the pending jobs in a Workflow.