Library

Jobs

SimpleWorkflows.Job — Type

Job(def; desc="", user="")

Create a simple job.

Arguments

def: A closure that encloses the job definition.

desc::String="": Describe briefly what this job does.

user::String="": Indicate who executes this job.

Examples

a = Job(() -> sleep(5); user="me", desc="Sleep for 5 seconds")
b = Job(() -> run(`pwd` & `ls`); user="me", desc="Run some commands")

SimpleWorkflows.@job — Macro

@job(ex, kwargs...)

Create a Job from an Expr, not a Function.

Examples

a = @job sleep(5) user="me" desc="Sleep for 5 seconds"
b = @job run(`pwd` & `ls`) user="me" desc="Run some commands"

SimpleWorkflows.getresult — Function

getresult(job::Job)

Get the running result of a Job.

The result is wrapped by a Some type. Use something to retrieve its value. If it is nothing, the Job is not finished.

SimpleWorkflows.getstatus — Method

getstatus(x::Job)

Get the current status of a Job.

SimpleWorkflows.ispending — Function

Test if the Job is still pending.

SimpleWorkflows.isrunning — Function

Test if the Job is running.

SimpleWorkflows.isexited — Function

Test if the Job has exited.

SimpleWorkflows.issucceeded — Function

Test if the Job was successfully run.

SimpleWorkflows.isfailed — Function

Test if the Job failed during running.

SimpleWorkflows.isinterrupted — Function

Test if the Job was interrupted during running.

SimpleWorkflows.createdtime — Function

Return the created time of a Job.

SimpleWorkflows.starttime — Function

starttime(job::Job)

Return the start time of a Job. Return nothing if it is still pending.

SimpleWorkflows.stoptime — Function

stoptime(job::Job)

Return the stop time of a Job. Return nothing if it has not exited.

SimpleWorkflows.elapsed — Function

elapsed(job::Job)

Return the elapsed time of a Job since it started running.

If nothing, the Job is still pending. If it is finished, return how long it took to complete.

SimpleWorkflows.description — Function

description(job::Job)

Return the description of a Job.

SimpleWorkflows.run! — Method

run!(job::Job; n=1, δt=1)

Run a Job with maximum n attempts, with each attempt separated by δt seconds.

SimpleWorkflows.interrupt! — Function

interrupt!(job::Job)

Manually interrupt a Job, works only if it is running.

SimpleWorkflows.queue — Function

queue(; sortby = :created_time)

Print all Jobs that are pending, running, or finished as a table.

Accpetable arguments for sortby are :created_time, :user, :start_time, :stop_time, :elapsed, :status, and :times.

SimpleWorkflows.query — Function

query(id::Integer)
query(ids::AbstractVector{<:Integer})

Query a specific (or a list of Jobs) by its (theirs) ID.

SimpleWorkflows.ntimes — Function

ntimes(id::Integer)
ntimes(job::Job)

Return how many times a Job has been rerun.

Workflows

SimpleWorkflows.Workflow — Type

Workflow(jobs, graph)

Create a Workflow from a list of Jobs and a graph representing their relations.

SimpleWorkflows.run! — Method

run!(wf::Workflow; n=5, δt=1, Δt=1, filename="saved.jld2")

Run a Workflow with maximum n attempts, with each attempt separated by Δt seconds.

Cool down for δt seconds after each Job in the Workflow. Save the tracking information to a file named saved.jld2.

SimpleWorkflows.getstatus — Method

getstatus(wf::Workflow)

Get the current status of each Job in a Workflow.

SimpleWorkflows.chain — Function

chain(x::Job, y::Job, z::Job...)

Chain multiple Jobs one after another.

SimpleWorkflows.:→ — Function

→(x, y)

Chain two Jobs.

SimpleWorkflows.:← — Function

←(y, x)

Chain two Jobs reversely.

SimpleWorkflows.thread — Function

thread(xs::AbstractVector{Job}, ys::AbstractVector{Job}, zs::AbstractVector{Job}...)

Chain multiple vectors of Jobs, each Job in xs has a corresponding Job in ys.`

SimpleWorkflows.:⇶ — Function

⇶(xs, ys)

Chain two vectors of Jobs.

SimpleWorkflows.:⬱ — Function

⬱(ys, xs)

Chain two vectors of Jobs reversely.

SimpleWorkflows.fork — Function

fork(x::Job, ys::AbstractVector{Job})
⇉(x, ys)

Attach a group of parallel Jobs (ys) to a single Job (x).

SimpleWorkflows.converge — Function

converge(xs::AbstractVector{Job}, y::Job)
⭃(xs, y)

Finish a group a parallel Jobs (xs) by a single Job (y).

SimpleWorkflows.spindle — Function

spindle(x::Job, ys::AbstractVector{Job}, z::Job)

Start from a Job (x), followed by a series of Jobs (ys), finished by a single Job (z).