Public API

SimpleWorkflows.Workflow — Type

Workflow(jobs::AbstractJob...)

Create a Workflow from a given series of AbstractJobs.

The list of AbstractJobs does not have to be complete, our algorithm will find all connected AbstractJobs automatically.

SimpleWorkflows.AutosaveWorkflow — Type

AutosaveWorkflow(path, jobs::AbstractJob...)

Create a AutosaveWorkflow from a given series of Jobs and a path.

When running, the status of the workflow will be automatically saved to path.

EasyJobsBase.run! — Function

run!(wf::Workflow; n=5, δt=1, Δt=1)

Run a Workflow with maximum n attempts, with each attempt separated by Δt seconds.

Cool down for δt seconds after each Job in the Workflow.

EasyJobsBase.getstatus — Function

getstatus(wf::AbstractWorkflow)

Get the current status of Jobs in a AbstractWorkflow.

EasyJobsBase.pendingjobs — Function

pendingjobs(wf::AbstractWorkflow)

Filter only the pending jobs in a Workflow.

EasyJobsBase.runningjobs — Function

runningjobs(wf::AbstractWorkflow)

Filter only the running jobs in a Workflow.

EasyJobsBase.exitedjobs — Function

exitedjobs(wf::AbstractWorkflow)

Filter only the exited jobs in a Workflow.

EasyJobsBase.succeededjobs — Function

succeededjobs(wf::AbstractWorkflow)

Filter only the succeeded jobs in a Workflow.

EasyJobsBase.failedjobs — Function

failedjobs(wf::AbstractWorkflow)

Filter only the failed jobs in a Workflow.

EasyJobsBase.interruptedjobs — Function

interruptedjobs(wf::AbstractWorkflow)

Filter only the interrupted jobs in a Workflow.