Public API

SimpleWorkflows.Workflow — Type

Workflow(jobs::AbstractJob...)

Create a Workflow from a given series of AbstractJobs.

The list of AbstractJobs does not have to be complete, our algorithm will find all connected AbstractJobs automatically.

EasyJobsBase.run! — Function

run!(wf::Workflow; maxattempts=5, interval=1, delay=0)

Run a Workflow with maximum number of attempts, with each attempt separated by a few seconds.

EasyJobsBase.execute! — Function

execute!(workflow::AbstractWorkflow, exec::Executor)

Executes the jobs from the workflow of the provided Executor instance.

The function will attempt to execute all the jobs up to exec.maxattempts times. If all jobs have succeeded, the function will stop immediately. Otherwise, it will wait for a duration equal to exec.interval before the next attempt.

EasyJobsBase.getstatus — Function

getstatus(wf::AbstractWorkflow)

Get the current status of jobs in a AbstractWorkflow as a graph.

SimpleWorkflows.liststatus — Function

liststatus(wf::AbstractWorkflow)

List the current status of jobs in a AbstractWorkflow as a vector.

See also getstatus.

EasyJobsBase.ispending — Function

ispending(wf::AbstractWorkflow)

Check if all jobs in the AbstractWorkflow are in a pending state.

Return true if all jobs are pending, otherwise, return false.

EasyJobsBase.isrunning — Function

isrunning(wf::AbstractWorkflow)

Check if any job in the AbstractWorkflow is currently running.

Return true if at least one job is running, otherwise, return false.

EasyJobsBase.isexited — Function

isexited(wf::AbstractWorkflow)

Check if all jobs in the AbstractWorkflow have exited.

Return true if all jobs have exited, otherwise, return false.

EasyJobsBase.issucceeded — Function

issucceeded(wf::AbstractWorkflow)

Check if all jobs in the AbstractWorkflow have successfully completed.

Return true if all jobs have succeeded, otherwise, return false.

EasyJobsBase.isfailed — Function

isfailed(wf::AbstractWorkflow)

Check if any job in the AbstractWorkflow has failed, given that all jobs have exited.

Return true if any job has failed after all jobs have exited, otherwise, return false.

EasyJobsBase.listpending — Function

listpending(wf::AbstractWorkflow)

Filter only the pending jobs in a Workflow.

EasyJobsBase.listrunning — Function

listrunning(wf::AbstractWorkflow)

Filter only the running jobs in a Workflow.

EasyJobsBase.listexited — Function

listexited(wf::AbstractWorkflow)

Filter only the exited jobs in a Workflow.

EasyJobsBase.listsucceeded — Function

listsucceeded(wf::AbstractWorkflow)

Filter only the succeeded jobs in a Workflow.

EasyJobsBase.listfailed — Function

listfailed(wf::AbstractWorkflow)

Filter only the failed jobs in a Workflow.