Public API

SimpleWorkflows.Workflow — Type

Workflow(jobs::AbstractJob...)

Create a Workflow from a given series of AbstractJobs.

The list of AbstractJobs does not have to be complete, our algorithm will find all connected AbstractJobs automatically.

EasyJobsBase.run! — Function

run!(wf::Workflow; n=5, δt=1, Δt=1)

Run a Workflow with maximum n attempts, with each attempt separated by Δt seconds.

Cool down for δt seconds after each Job in the Workflow.

EasyJobsBase.execute! — Function

execute!(exec::Executor)

Executes the jobs from the workflow of the provided Executor instance.

The function will attempt to execute all the jobs up to maxattempts times. If all jobs have succeeded, the function will stop immediately. Otherwise, it will wait for a given interval before the next attempt.

EasyJobsBase.getstatus — Function

getstatus(wf::AbstractWorkflow)

Get the current status of Jobs in a AbstractWorkflow as a graph.

SimpleWorkflows.liststatus — Function

liststatus(wf::AbstractWorkflow)

List the current status of Jobs in a AbstractWorkflow as a vector.

See also getstatus.

Missing docstring.

Missing docstring for pendingjobs. Check Documenter's build log for details.

Missing docstring.

Missing docstring for runningjobs. Check Documenter's build log for details.

Missing docstring.

Missing docstring for exitedjobs. Check Documenter's build log for details.

Missing docstring.

Missing docstring for succeededjobs. Check Documenter's build log for details.

Missing docstring.

Missing docstring for failedjobs. Check Documenter's build log for details.

Missing docstring.

Missing docstring for interruptedjobs. Check Documenter's build log for details.